remove reference to framework support
building universal frameworks is still a bit of a hacky process
involving lipo and manual code-signing.

going to wait to implement this (if at all) until more is known about
how to integrate frameworks into your development workflow in such a
way that it is possible to develop for the device and the simulator at
the same time.

for now, going to focus on delivering cocoapods functionality